test(TaskGraph): Test that TaskGraph runs nodes once
Although runNode > getNodeInputs > getNodeOutput form a mutually recursive chain, which could
conceivably prevent the result of a node run() from being stored before it is requested again,
.run() calls are scheduled via the microtask queue, as they are in the then() blocks of the
Promise.all() request for a node's inputs. Once a node with no dependencies is reached, its run() is
scheduled on the microtask queue, the promise containing its result is returned, and the
Promise.all() that requested it can repeat this for its remaining dependencies; ultimately, the
correct partial order is constructed during a single pass of the JS event queue, and during the
dequeueing of the microtask queue, the actual runs begin, proceeding in valid dependency (partial)
order.